Refactor mute handling by replacing mute_check.py with mute_helper.py
- Deleted the obsolete mute_check.py file and introduced mute_helper.py to enhance the organization of mute workflows.
- Updated import statements across various scripts to reflect the new module structure under the `mute` directory.
- Adjusted the update_muted_ya.yml workflow to utilize the new mute_helper.py for improved clarity and maintainability.
